feat: 에러 상황 추가
- 메모리 캐시, 디스크 캐시, 이미지 다운로더 각각에 대한 에러 설정
